created: SigningConfig.kt and extension about signing config added to it
